+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KGQB2J5NP8YM1PMZ7PQWE6WZ"
+        "01KGRPZAF2J0TXCZ4DAMJRFVDK"
     )
     r"""The unique identifier of the data source"""